+++ adalflow/adalflow/components/model_client/openai_client.py
@@ -367,7 +367,7 @@ class OpenAIClient(ModelClient):
         else:
             raise ValueError(f"model_type {model_type} is not supported")
 
-        print(f"final_model_kwargs: {final_model_kwargs}")
+        # print(f"final_model_kwargs: {final_model_kwargs}")
 
         return final_model_kwargs
 
@@ -549,14 +549,20 @@ if __name__ == "__main__":
     log = get_logger(level="DEBUG")
 
     setup_env()
-    prompt_kwargs = {"input_str": "What is the meaning of life?"}
+    # prompt_kwargs = {"input_str": "What is the meaning of life?"}
 
-    gen = Generator(
-        model_client=OpenAIClient(),
-        model_kwargs={"model": "gpt-3.5-turbo", "stream": True},
-    )
-    gen_response = gen(prompt_kwargs)
-    print(f"gen_response: {gen_response}")
+    # gen = Generator(
+    #     model_client=OpenAIClient(),
+    #     model_kwargs={"model": "gpt-3.5-turbo", "stream": True},
+    # )
+    # gen_response = gen(prompt_kwargs)
+    # print(f"gen_response: {gen_response}")
 
-    for genout in gen_response.data:
-        print(f"genout: {genout}")
+    # for genout in gen_response.data:
+    #     print(f"genout: {genout}")
+
+    # test that to_dict and from_dict works
+    model_client = OpenAIClient()
+    model_client_dict = model_client.to_dict()
+    from_dict_model_client = OpenAIClient.from_dict(model_client_dict)
+    assert model_client_dict == from_dict_model_client.to_dict()